Exclude class to files
