+++ src/Lekarz.cpp
@@ -1,3 +1,5 @@
+#include <iostream>
+
 #include "Lekarz.h"
 
 Lekarz::Lekarz(string imie, string nazwisko, string specjalizacja, int ID_lekarza, int gabinet) {
@@ -6,17 +8,13 @@ Lekarz::Lekarz(string imie, string nazwisko, string specjalizacja, int ID_lekarz
     this->specjalizacja = specjalizacja;
     this->ID_lekarza = ID_lekarza;
     this->gabinet = gabinet;
-}
+};
 
-void Lekarz::wyswietl() {
+void Lekarz::wyswietl(){
     cout << "Imie: " << imie << endl;
     cout << "Nazwisko: " << nazwisko << endl;
     cout << "Specjalizacja: " << specjalizacja << endl;
     cout << "ID lekarza: " << ID_lekarza << endl;
     cout << "Gabinet: " << gabinet << endl;
-}
-
-// Lekarz::~Lekarz() {
-//     cout << "Destruktor klasy Lekarz" << endl;
-// }
+};
