+++ src/Pacjent.h
@@ -3,11 +3,10 @@
 
 #include <iostream>
 #include <string>
-#include "Wizyta.h"
 
 using namespace std;
 
-class Pacjent : public Wizyta{
+class Pacjent{
 private:
     string imie;
     string nazwisko;
@@ -19,7 +18,7 @@ public:
     void wyswietl();
 };
 
-class KartaPacjenta : public Pacjent {
+class KartaPacjenta : public Pacjent{
 private:
     Pacjent pacjent;
     string data_urodzenia;
@@ -27,21 +26,9 @@ private:
     string m_urodzenia;
     int nr_telefonu;
 public:
-    KartaPacjenta(string imie, string nazwisko, unsigned long long pesel, string data_urodzenia, string m_zamieszkania, string m_urodzenia, int nr_telefonu);
+    KartaPacjenta(){}; // konstruktor domyślny
+    KartaPacjenta(Pacjent pacjent, string data_urodzenia, string m_zamieszkania, string m_urodzenia, int nr_telefonu);
     void wyswietl();
-    int umowWizyte(Wizyta wizyta){
-
-        cout << "Wybierz date wizyty z dostępnych terminów" << endl;
-        pobierzDatyWizyt(); 
-        Wizyta wizyta = {
-            data_wizyty,
-            godzina_wizyty,
-            lekarz,
-            opis_wizyty
-            };
-        }
-
-    }
 };
 
 #endif
