+++ src/Wizyta.h
@@ -3,25 +3,25 @@
 
 #include <iostream>
 #include <string>
+#include "helpers.h"
 #include "Pacjent.h"
 #include "Lekarz.h"
+
 
 using namespace std;
 
 
-class Wizyta : public Pacjent, public Lekarz{
+class Wizyta{
 private:
-        string termin;
-        Lekarz lekarz;
-        Pacjent pacjent;
-        char status_wizyty;
-        int koszt;
-    public:
-        Wizyta(){}; // konstruktor domyślny
-        Wizyta(string imie, string nazwisko, unsigned long long pesel, string data_urodzenia, string m_zamieszkania, string m_urodzenia,
-               int nr_telefonu, string imie_lekarza, string nazwisko_lekarza, string specjalizacja, int ID_lekarza, int gabinet,
-               string termin, char status_wizyty, int koszt){};
-               void wyswietl();
+    string termin;
+    Lekarz lekarz;
+    Pacjent pacjent;
+    StatusWizyty status_wizyty;
+    int koszt;
+public:
+    Wizyta(){}; // konstruktor domyślny
+    Wizyta(Pacjent pacjent, Lekarz lekarz, string termin, StatusWizyty status_wizyty, int koszt);
+    void wyswietl();
 };
 
 class Recepta : public Wizyta{
@@ -32,21 +32,17 @@ private:
     string dawkowanie;
 public:
     Recepta(){}; // konstruktor domyślny
-    Recepta(string imie, string nazwisko, unsigned long long pesel, string data_urodzenia, string m_zamieszkania, 
-    string m_urodzenia, int nr_telefonu, string imie_lekarza, string nazwisko_lekarza, string specjalizacja, 
-    int ID_lekarza, int gabinet, string termin, char status_wizyty, int koszt, string lek, int ilosc, string dawkowanie);
+    Recepta(Wizyta wizyta, string lek, int ilosc, string dawkowanie);
     void wyswietl();
 };
 
-class Skierowanie : public Wizyta{
+class Skierowanie{
 private:
     Wizyta wizyta;
     string badanie;
 public:
     Skierowanie(){}; // konstruktor domyślny
-    Skierowanie(string imie, string nazwisko, unsigned long long pesel, string data_urodzenia, string m_zamieszkania,
-    string m_urodzenia, int nr_telefonu, string imie_lekarza, string nazwisko_lekarza, string specjalizacja,
-    int ID_lekarza, int gabinet, string termin, char status_wizyty, int koszt, string badanie);
+    Skierowanie(Wizyta wizyta, string badanie);
     void wyswietl();
 };
